% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_functions.R
\name{bonf_t}
\alias{bonf_t}
\title{Naive method}
\usage{
bonf_t(ds, N.outcomes)
}
\arguments{
\item{ds}{the simulated data and the total number of outcomes}
}
\value{
P-values from each method
}
\description{
This function uses prop.test and implements Bonferroni test
}
