% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_functions.R
\name{compute_power}
\alias{compute_power}
\title{Power and type I error}
\usage{
compute_power(l.result, N.sim)
}
\arguments{
\item{l.result}{p-values from each dataset and method}
}
\value{
Power and type I error from each method
}
\description{
This function computes the power and the type I error from each method
}
