% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_functions.R
\name{compute_pvalue_obs}
\alias{compute_pvalue_obs}
\title{P-value real data}
\usage{
compute_pvalue_obs(ds, permute)
}
\arguments{
\item{ds}{the real-world data and the matrix with the permuted indeces specified}
}
\value{
P-values from each method
}
\description{
This function computes the p-value for real-world data using all methods
}
