% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_functions.R
\name{compute_uncertainty_boot}
\alias{compute_uncertainty_boot}
\title{Uncertainty for bootstrapping}
\usage{
compute_uncertainty_boot(ds)
}
\arguments{
\item{ds}{the simulated data}
}
\value{
P-values from each method
}
\description{
This function computes the uncertainty around the bootstrapped test statistics
}
