% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_functions.R
\name{gen_data_corr}
\alias{gen_data_corr}
\title{Generate simulated data}
\usage{
gen_data_corr(RR, prop.outcome, N1, N2, N.outcomes, cor)
}
\arguments{
\item{RR}{Params are risk ratio, incidence of outcomes, N subjects in both groups, number of outcomes and correlation among outcomes}
}
\value{
The simulated data
}
\description{
This function creates simulated correlated data. It creates both the treatment and control groups
with N.outcomes endpoints with certain correlation speficied by the user
}
