% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_functions.R
\name{load_result}
\alias{load_result}
\title{Load results from permutation + naive methods}
\usage{
load_result(pattern)
}
\arguments{
\item{pattern}{the file path}
}
\value{
List of the results
}
\description{
This function extracts the results from naive and permutation methods
}
