% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_functions.R
\name{load_result_ind}
\alias{load_result_ind}
\title{Load results from independent outcomes}
\usage{
load_result_ind(pattern)
}
\arguments{
\item{pattern}{the file path}
}
\value{
List of the results
}
\description{
This function extracts the results from outcomes treated as independent
}
