% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_functions.R
\name{perm.fun}
\alias{perm.fun}
\title{Run permutation and naive functions}
\usage{
perm.fun(ds, permute)
}
\arguments{
\item{ds}{the simulated data and a matrix permute to indicate the permuted indeces}
}
\value{
P-values from each method
}
\description{
This function takes the simulated data and computes the test statistics using permutation approaches
}
