% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_functions.R
\name{prop_test}
\alias{prop_test}
\title{Proportion test}
\usage{
prop_test(N.vax.outcomes, N.unvax.outcomes, N.vax, N.unvax, N.outcomes)
}
\arguments{
\item{N.vax.outcomes}{the number of cases in the treatment group, in the control group, the total number of people in treatment and control
groups and the total number of outcomes}
}
\value{
P-values from prop test
}
\description{
This function computes prop_test and stores the p-values
}
