% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_functions.R
\name{RR_naive}
\alias{RR_naive}
\title{Naive method}
\usage{
RR_naive(ds, N.outcomes)
}
\arguments{
\item{ds}{the simulated data}
}
\value{
P-values from each method
}
\description{
This function computes the p-value for each independent outcome and selects the smaller p-value
}
