% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_functions.R
\name{RR_naive_obs}
\alias{RR_naive_obs}
\title{Naive method for observed data}
\usage{
RR_naive_obs(ds)
}
\arguments{
\item{ds}{the simulated data}
}
\value{
P-values from each method
}
\description{
This function computes the p-value for each independent outcome and selects the smaller p-value using observed data
}
