% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_functions.R
\name{set_setting}
\alias{set_setting}
\title{The user can specify the settings that will be run. Either the user specifies a setting based on real-world clinical trials or uses custom parameters}
\usage{
set_setting(
  setting,
  N.sim = FALSE,
  N.outcomes = FALSE,
  RR = FALSE,
  prop.outcome = FALSE,
  N1 = FALSE,
  N2 = FALSE
)
}
\arguments{
\item{setting}{name of the setting to be simulated, if "custom", the user can specify the number of simulated datasets, number of outcomes, risk ratio, incidence of the outcomes
and number of individuals in control and treatment groups}
}
\value{
This function returns the setting's characteristics that will be the input in the main function
}
\description{
The user can specify the settings that will be run. Either the user specifies a setting based on real-world clinical trials or uses custom parameters
}
