% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_functions.R
\name{test_statistics_boot}
\alias{test_statistics_boot}
\title{Boostrapping method}
\usage{
test_statistics_boot(ds)
}
\arguments{
\item{ds}{the real-world data}
}
\value{
Bootstrapped test statistics
}
\description{
This function computes the bootstrapping for each permutation method
}
