% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_functions.R
\name{test_statistics_obs}
\alias{test_statistics_obs}
\title{Test statistics observed data}
\usage{
test_statistics_obs(ds)
}
\arguments{
\item{ds}{the simulated data}
}
\value{
Observed test statistics
}
\description{
This function computes the test statistcs of the observed data
}
