% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_functions.R
\name{load_result}
\alias{load_result}
\title{Plotting function}
\usage{
load_result(path, pattern)
}
\arguments{
\item{path}{where the results are stored, and specific pattern to load correct results}
}
\value{
A dataframe for plotting
}
\description{
This function manipulates the results for plotting
}
