% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_functions.R
\name{main_run}
\alias{main_run}
\title{Main function}
\usage{
main_run(N.sim, N.outcomes, RR, prop.outcome, N1, N2, cor)
}
\arguments{
\item{N.sim}{number of datasets, number of outcomes, RR, incidence, subjects in each category and correlation among outcomes}
}
\value{
Final results
}
\description{
This function creates the simulated data for each dataset and computes power and type I error
}
