% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_functions.R
\name{set_setting}
\alias{set_setting}
\title{Function to decide which setting to run, here you specify}
\usage{
set_setting(
  setting,
  N.sim = FALSE,
  N.outcomes = FALSE,
  RR = FALSE,
  prop.outcome = FALSE,
  N1 = FALSE,
  N2 = FALSE
)
}
\arguments{
\item{setting}{name of the setting to be simulated, if "custom", the user can specify the number of simulated datasets, number of outcomes, risk ratio, incidence of the outcomes
and number of individuals in control and treatment groups}
}
\value{
setting characteristics
}
\description{
This function specifies settings based on real-world clinical trials or uses custom parameters that the user can input
}
